Fixed issue with turret_transform being off.
Essentially, it was a red-herring. Since mispredictions of turret rotation won't
cause a rollback, the turret_transform in the state won't always get updated.
Turret rotation is really input, not state - but we need to have it in state
to do interpolation! So, added the ability to have state properties that aren't
used in calculating the hash, but only interpolation. It's using the same
mechanism (and code!) as we use for excluding certain input properties from
the hash calculation.

This will be useful for when implement production checks for matching state.